pyproject.toml
[tool.poetry]
name = "depcheck"
version = "0.1.0"
description = ""
authors = ["Team Simplex <team.simplex@flixbus.com>"]
maintainers = ["Team Simplex <team.simplex@flixbus.com>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydeps = "^1.9.13"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pydeps = "^1.9.13"
PyYAML = "^5.4.1"
pytest = "^6.2.2"
pyre-check = "^0.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
